+++ FSCalendar/FSCalendarExtensions.m
@@ -168,7 +168,7 @@
     if (!week) return nil;
     NSDateComponents *weekdayComponents = [self components:NSCalendarUnitWeekday fromDate:week];
     NSDateComponents *components = self.fs_privateComponents;
-    components.day = - (weekdayComponents.weekday - self.firstWeekday);
+    components.day = - (weekdayComponents.weekday - (NSInteger)self.firstWeekday);
     components.day = (components.day-7) % 7;
     NSDate *firstDayOfWeek = [self dateByAddingComponents:components toDate:week options:0];
     firstDayOfWeek = [self dateBySettingHour:0 minute:0 second:0 ofDate:firstDayOfWeek options:0];
@@ -181,7 +181,7 @@
     if (!week) return nil;
     NSDateComponents *weekdayComponents = [self components:NSCalendarUnitWeekday fromDate:week];
     NSDateComponents *components = self.fs_privateComponents;
-    components.day = - (weekdayComponents.weekday - self.firstWeekday);
+    components.day = - (weekdayComponents.weekday - (NSInteger)self.firstWeekday);
     components.day = (components.day-7) % 7 + 6;
     NSDate *lastDayOfWeek = [self dateByAddingComponents:components toDate:week options:0];
     lastDayOfWeek = [self dateBySettingHour:0 minute:0 second:0 ofDate:lastDayOfWeek options:0];
